# [SofaSimpleFem] Simplify bloc-based optimization