--- a/SofaKernel/modules/SofaDefaultType/src/sofa/defaulttype/BaseMatrix.cpp
+++ b/SofaKernel/modules/SofaDefaultType/src/sofa/defaulttype/BaseMatrix.cpp
@@ -111,31 +111,31 @@ struct BaseMatrixLinearOpMV_BlockDiagonal
 };
 
 ///Adding values from a 3x3d matrix this function may be overload to obtain better performances
-void BaseMatrix::add(Index _i, Index _j, const type::Mat3x3d & _M) {
+void BaseMatrix::add(Index row, Index col, const type::Mat3x3d & _M) {
     for (unsigned i=0;i<3;i++)
         for (unsigned j=0;j<3;j++)
-            add(_i+i,_j+j,_M[i][j]);
+            add(row + i, col + j, _M[i][j]);
 }
 
 ///Adding values from a 3x3f matrix this function may be overload to obtain better performances
-void BaseMatrix::add(Index _i, Index _j, const type::Mat3x3f & _M) {
+void BaseMatrix::add(Index row, Index col, const type::Mat3x3f & _M) {
     for (unsigned i=0;i<3;i++)
         for (unsigned j=0;j<3;j++)
-            add(_i+i,_j+j,_M[i][j]);
+            add(row + i, col + j, _M[i][j]);
 }
 
 ///Adding values from a 2x2d matrix this function may be overload to obtain better performances
-void BaseMatrix::add(Index _i, Index _j, const type::Mat2x2d & _M) {
+void BaseMatrix::add(Index row, Index col, const type::Mat2x2d & _M) {
     for (unsigned i=0;i<2;i++)
         for (unsigned j=0;j<2;j++)
-            add(_i+i,_j+j,_M[i][j]);
+            add(row + i, col + j, _M[i][j]);
 }
 
 ///Adding values from a 2x2f matrix this function may be overload to obtain better performances
-void BaseMatrix::add(Index _i, Index _j, const type::Mat2x2f & _M) {
+void BaseMatrix::add(Index row, Index col, const type::Mat2x2f & _M) {
     for (unsigned i=0;i<2;i++)
         for (unsigned j=0;j<2;j++)
-            add(_i+i,_j+j,_M[i][j]);
+            add(row + i, col + j, _M[i][j]);
 }