--- a/SofaKernel/modules/SofaSimpleFem/src/SofaSimpleFem/HexahedronFEMForceField.inl
+++ b/SofaKernel/modules/SofaSimpleFem/src/SofaSimpleFem/HexahedronFEMForceField.inl
@@ -1124,80 +1124,32 @@ void HexahedronFEMForceField<DataTypes>::addKToMatrix(const core::MechanicalPara
     sofa::core::behavior::MultiMatrixAccessor::MatrixRef r = matrix->getMatrix(this->mstate);
     const Real kFactor = (Real)sofa::core::mechanicalparams::kFactorIncludingRayleighDamping(mparams, this->rayleighStiffness.getValue());
 
-    if (auto* crsmat_d = dynamic_cast<sofa::component::linearsolver::CompressedRowSparseMatrix<type::Mat<3,3,double> > * >(r.matrix))
+    sofa::Index e { 0 }; //index of the element in the topology
+
+    const auto& stiffnesses = _elementStiffnesses.getValue();
+    const auto* indexedElements = this->getIndexedElements();
+
+    for (const auto& element : *indexedElements)
     {
-        addKToBlocMatrix<double>(crsmat_d, kFactor, r.offset);
-    }
-    else if (auto* crsmat_f = dynamic_cast<sofa::component::linearsolver::CompressedRowSparseMatrix<type::Mat<3,3,float> > * >(r.matrix))
-    {
-        addKToBlocMatrix<float>(crsmat_f, kFactor, r.offset);
-    }
-    else
-    {
-        int e; //index of the element in the topology
-
-        typename VecElement::const_iterator it;
-        typename Element::size_type n1, n2;
-
-        Index node1, node2;
-
-        for(it = this->getIndexedElements()->begin(), e=0 ; it != this->getIndexedElements()->end() ; ++it,++e)
-        {
-            const ElementStiffness &Ke = _elementStiffnesses.getValue()[e];
-            const Transformation Rot = getElementRotation(e);
-
-            // find index of node 1
-            for (n1 = 0; n1 < Element::size(); n1++)
-            {
-                node1 = (*it)[n1];
-                // find index of node 2
-                for (n2 = 0; n2 < Element::size(); n2++)
-                {
-                    node2 = (*it)[n2];
-
-                    const Mat33 tmp = Rot.multTranspose( Mat33(
-                            Coord(Ke[3*n1+0][3*n2+0],Ke[3*n1+0][3*n2+1],Ke[3*n1+0][3*n2+2]),
-                            Coord(Ke[3*n1+1][3*n2+0],Ke[3*n1+1][3*n2+1],Ke[3*n1+1][3*n2+2]),
-                            Coord(Ke[3*n1+2][3*n2+0],Ke[3*n1+2][3*n2+1],Ke[3*n1+2][3*n2+2])) ) * Rot;
-                    r.matrix->add(r.offset + 3 * node1, r.offset + 3 * node2, tmp * (-kFactor));
-                }
-            }
-        }
-    }
-}
-
-template<class DataTypes>
-template<class BlocReal>
-void HexahedronFEMForceField<DataTypes>::addKToBlocMatrix(
-        sofa::component::linearsolver::CompressedRowSparseMatrix<type::Mat<3, 3, BlocReal>, type::vector<type::Mat<3, 3, BlocReal> >, type::vector<sofa::Index> >* crsmat,
-        SReal k, unsigned int& offset)
-{
-    int e; //index of the element in the topology
-
-    typename VecElement::const_iterator it;
-    typename Element::size_type n1, n2;
-
-    Index node1, node2;
-
-    for(it = this->getIndexedElements()->begin(), e=0 ; it != this->getIndexedElements()->end() ; ++it,++e)
-    {
-        const ElementStiffness &Ke = _elementStiffnesses.getValue()[e];
+        const ElementStiffness &Ke = stiffnesses[e];
         const Transformation Rot = getElementRotation(e);
+        e++;
 
         // find index of node 1
-        for (n1 = 0; n1 < Element::size(); n1++)
+        for (Element::size_type n1 = 0; n1 < Element::size(); n1++)
         {
-            node1 = (*it)[n1];
+            const auto node1 = element[n1];
             // find index of node 2
-            for (n2 = 0; n2 < Element::size(); n2++)
+            for (Element::size_type n2 = 0; n2 < Element::size(); n2++)
             {
-                node2 = (*it)[n2];
+                const auto node2 = element[n2];
 
                 const Mat33 tmp = Rot.multTranspose( Mat33(
                         Coord(Ke[3*n1+0][3*n2+0],Ke[3*n1+0][3*n2+1],Ke[3*n1+0][3*n2+2]),
                         Coord(Ke[3*n1+1][3*n2+0],Ke[3*n1+1][3*n2+1],Ke[3*n1+1][3*n2+2]),
                         Coord(Ke[3*n1+2][3*n2+0],Ke[3*n1+2][3*n2+1],Ke[3*n1+2][3*n2+2])) ) * Rot;
-                *crsmat->wbloc(offset / 3 + node1, offset / 3 + node2, true) += tmp * (-k);
+
+                r.matrix->add( r.offset + 3 * node1, r.offset + 3 * node2, tmp * (-kFactor));
             }
         }
     }