--- a/modules/SofaSparseSolver/examples/FEM_SparseLDLSolver.scn
+++ b/modules/SofaSparseSolver/examples/FEM_SparseLDLSolver.scn
@@ -10,7 +10,9 @@
     <Node name="M1">
         <EulerImplicitSolver name="odesolver" printLog="false"  rayleighStiffness="0.1" rayleighMass="0.1" />
         <SparseLDLSolver printLog="false" template="CompressedRowSparseMatrixMat3x3d"/>
-<!--        <SparseLDLSolver printLog="false" template="CompressedRowSparseMatrixd"/>-->
+        <!-- You can assemble the matrix using scalar instead of blocks by changing the SparseLDLSolver template as follows:
+            <SparseLDLSolver printLog="false" template="CompressedRowSparseMatrixd"/>
+        -->
         <MechanicalObject />
         <UniformMass vertexMass="1" />
         <RegularGridTopology nx="4" ny="4" nz="20" xmin="-9" xmax="-6" ymin="0" ymax="3" zmin="0" zmax="19" />