@fredroy 2021-08-04
Needs rebase when #2280 is merged